+++ modules/tinymce/Gruntfile.js
@@ -7,7 +7,7 @@ let swag = require('@ephox/swag');
 
 let plugins = [
   'advlist', 'anchor', 'autolink', 'autoresize', 'autosave', 'charmap', 'code', 'codesample',
-  'directionality', 'emoticons', 'help', 'fullscreen', 'hr', 'image', 'imagetools', 'importcss',
+  'directionality', 'emoticons', 'help', 'fullscreen', 'hr', 'image', 'importcss',
   'insertdatetime', 'link', 'lists', 'media', 'nonbreaking', 'noneditable', 'pagebreak', 'paste',
   'preview', 'print', 'save', 'searchreplace', 'tabfocus', 'table', 'template', 'textpattern',
   'toc', 'visualblocks', 'visualchars', 'wordcount', 'quickbars'