+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/contextmenu/MobileContextMenuTest.ts
@@ -7,7 +7,6 @@ import { TinyDom, TinyHooks, TinySelections, TinyUiActions } from '@ephox/wrap-m
 
 import Editor from 'tinymce/core/api/Editor';
 import ImagePlugin from 'tinymce/plugins/image/Plugin';
-import ImageToolsPlugin from 'tinymce/plugins/imagetools/Plugin';
 import LinkPlugin from 'tinymce/plugins/link/Plugin';
 import TablePlugin from 'tinymce/plugins/table/Plugin';
 
@@ -37,12 +36,12 @@ describe('browser.tinymce.themes.silver.editor.contextmenu.MobileContextMenuTest
   });
 
   const hook = TinyHooks.bddSetupLight<Editor>({
-    plugins: 'image imagetools link table',
+    plugins: 'image link table',
     toolbar: 'image editimage link table',
     indent: false,
     base_url: '/project/tinymce/js/tinymce',
     image_caption: true
-  }, [ ImagePlugin, ImageToolsPlugin, LinkPlugin, TablePlugin ], true);
+  }, [ ImagePlugin, LinkPlugin, TablePlugin ], true);
 
   const pOpenContextMenu = async (editor: Editor, target: string) => {
     const targetElem = UiFinder.findIn(TinyDom.body(editor), target).getOrDie();
@@ -57,8 +56,6 @@ describe('browser.tinymce.themes.silver.editor.contextmenu.MobileContextMenuTest
   const pressDownArrowKey = () => Keyboard.activeKeydown(SugarDocument.getDocument(), Keys.down());
   const pressEnterKey = () => Keyboard.activeKeydown(SugarDocument.getDocument(), Keys.enter());
   const pressEscKey = () => Keyboard.activeKeydown(SugarDocument.getDocument(), Keys.escape());
-
-  const repeatDownArrowKey = (index: number) => Arr.range(index, pressDownArrowKey);
 
   const tableHtml = '<table style="width: 100%;">' +
   '<tbody>' +
@@ -170,7 +167,6 @@ describe('browser.tinymce.themes.silver.editor.contextmenu.MobileContextMenuTest
     assertMenuItems([
       selectors.link,
       selectors.image,
-      selectors.editimage,
       selectors.cell,
       selectors.row,
       selectors.column,
@@ -178,12 +174,6 @@ describe('browser.tinymce.themes.silver.editor.contextmenu.MobileContextMenuTest
       selectors.deletetable
     ]);
     FocusTools.setFocus(SugarBody.body(), selectors.image);
-    pressEnterKey();
-    await pWaitForAndCloseDialog(editor);
-    await pOpenContextMenu(editor, 'img');
-    // Navigate to the "Image tools" menu item
-    FocusTools.setFocus(SugarBody.body(), selectors.link);
-    repeatDownArrowKey(2);
     pressEnterKey();
     await pWaitForAndCloseDialog(editor);
   });