+++ modules/tinymce/CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The legacy `mobile` theme has been removed #TINY-7832
 - Removed support for Microsoft Internet Explorer 11 #TINY-8194
 - Removed the deprecated `fullpage`, `spellchecker`, `bbcode`, `legacyoutput`, `colorpicker`, `contextmenu` and `textcolor` plugins #TINY-8192
+- Removed support for Word from the `paste` plugin #TINY-7493
+- Removed the `imagetools` plugin, which is now classified as a Premium plugin #TINY-8209
 
 ## 5.10.1 - 2021-11-03